api/requirements.txt
altair==5.3.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
exceptiongroup==1.1.2
fastapi==0.100.1
filelock==3.12.2
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
requests==2.31.0
rich==13.5.1
rpds-py==0.9.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.20.0
zipp==3.16.2
transformers 
sentencepiece
